Switch middleware to exceptions, move to folder
