+++ packages/api-gateway/src/modules/user/user.module.ts
@@ -12,7 +12,7 @@ import { UserController } from './user.controller';
 import { UserProto, UserProtoFile, JwtProto, JwtProtoFile } from 'juno-proto';
 
 const { JWT_SERVICE_NAME, JUNO_JWT_PACKAGE_NAME } = JwtProto;
-import { CredentialsMiddleware } from 'src/credentials.middleware';
+import { CredentialsMiddleware } from 'src/middleware/credentials.middleware';
 
 const { USER_SERVICE_NAME, USER_AUTH_SERVICE_NAME, JUNO_USER_PACKAGE_NAME } =
   UserProto;
@@ -60,6 +60,9 @@ export class UserModule implements NestModule {
       .forRoutes({ path: 'user/id/:id/project', method: RequestMethod.PUT });
     consumer
       .apply(CredentialsMiddleware)
-      .forRoutes({ path: 'user', method: RequestMethod.POST });
+      .forRoutes(
+        { path: 'user', method: RequestMethod.POST },
+        { path: 'user/type', method: RequestMethod.POST },
+      );
   }
 }
